[ALPS04140152] wmt driver: fix wmt test case
if android OS is 32bit and kernel is 64bit,
some sizeof data type is different.
And it will make wmt test case fail.

MTK-Commit-Id: b871dd55e265d9d1a97e920af83575418b271eeb

CR-Id: ALPS04140152
Feature: [Module]WMT Driver

+++ common_main/linux/wmt_dbg.c
@@ -1020,7 +1020,7 @@ INT32 wmt_dbg_ut_test(INT32 par1, INT32 par2, INT32 par3)
 
 #if CFG_CORE_INTERNAL_TXRX
 struct lpbk_package {
-	LONG payload_length;
+	UINT32 payload_length;
 	UINT8 out_payload[2048];
 	UINT8 in_payload[2048];
 };
